Updated project module
