+++ web/modules/custom/hoeringsportal_project/hoeringsportal_project.services.yml
@@ -3,3 +3,5 @@ services:
     autowire: true
 
   Drupal\hoeringsportal_project\Hook\ProjectHooks:
+
+  Drupal\hoeringsportal_project\Helper\Helper:
